Radically simplify checksum function
I came across someone talking about `partition-all` with a `step` of
`1`, and playing around with that reminded me that `partition` will give
me the grouping I need for the checksum function, which then makes it
trivial to wrap the whole thing in a thread-last form, drastically
simplifying the code, while also making it more idiomatic.

This is very satisfying.