Update docstrings for three core functions
I wanted to ensure each function has a docstring, so I added them in.

I also moved `checksum-entry` to be a private function, as you're really
meant to use `valid-entry?`, but I didn't want to have both the checksum
math and the validation in the same function. I can't quite say why, but
it felt in my gut like a worthwhile distinction.