Add a namespace docstring, and mark most fn's private
I like the `^:private` metadata as a helpful flag for people coming to a
namespace to understand what kind of functionality should be available
for use in other places in a project. As such, I've marked every
function that isn't meant for public consumption as private.

For the tests that use these newly-private functions, I've used the `#'`
reader macro to get their var. This is a practice I've picked up
elsewhere that allows you to have the clean delineation of public use
while still testing your private functions.